pyproject.toml
[project]
name = "adk-imagen-veo"
version = "0.1.0"
description = "ADK Agent for Image and Video Generation."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-adk==1.0.0",
    "google-genai==1.16.1",
]
